pyproject.toml
[project]
name = "remarcable"
version = "0.1.0"
description = "Django project for Remarcable"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref==3.11.1",
    "django==5.2.13",
    "pytest-django>=4.12.0",
    "ruff>=0.15.12",
    "sqlparse==0.5.5",
    "typing-extensions==4.15.0",
]
